moving into chat and adjusting the chat history
